new: add support of ``LDAP_ADMIN_BIND_PWD_FILE``
In original code, it is the entrypoint ensuring the support of all
``*_FILE`` environment variables. If not using the current docker
container, the support for all ``*_FILE`` environment file is lost.

The most important variable that could need that is
``LDAP_ADMIN_BIND_PWD_FILE`` and this is why the implementation is
added in the php code.